requirements.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.6
django-cors-headers==3.13.0
django-import-export==3.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
et-xmlfile==1.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
PyJWT==2.4.0
pytz==2022.1
PyYAML==6.0
sqlparse==0.4.2
tablib==3.3.0
xlrd==2.0.1
xlwt==1.3.0
